+nightly fmt

+++ crates/router/src/core/admin.rs
@@ -2,11 +2,11 @@ use api_models::admin::PrimaryBusinessDetails;
 use common_utils::{
     crypto::{generate_cryptographically_secure_random_string, OptionalSecretValue},
     date_time,
-    ext_traits::{ValueExt, Encode},
+    ext_traits::{Encode, ValueExt},
 };
-use error_stack::{report, FutureExt, ResultExt, IntoReport};
-use masking::{Secret, PeekInterface, ExposeInterface};
-use storage_models::{enums, configs::ConfigNew};
+use error_stack::{report, FutureExt, IntoReport, ResultExt};
+use masking::{ExposeInterface, PeekInterface, Secret};
+use storage_models::{configs::ConfigNew, enums};
 use uuid::Uuid;
 
 use crate::{
@@ -468,12 +468,14 @@ pub async fn create_payment_connector(
         created_at: common_utils::date_time::now(),
         modified_at: common_utils::date_time::now(),
         id: None,
-        connector_webhook_details: match req.connector_webhook_details{
+        connector_webhook_details: match req.connector_webhook_details {
             Some(connector_webhook_details) => {
-                Encode::<api_models::admin::MerchantConnectorWebhookDetails>::encode_to_value(&connector_webhook_details)
-                    .change_context(errors::ApiErrorResponse::InternalServerError)
-                    .map(|value| Some(Secret::new(value)))?
-            },
+                Encode::<api_models::admin::MerchantConnectorWebhookDetails>::encode_to_value(
+                    &connector_webhook_details,
+                )
+                .change_context(errors::ApiErrorResponse::InternalServerError)
+                .map(|value| Some(Secret::new(value)))?
+            }
             None => None,
         },
     };
@@ -486,27 +488,23 @@ pub async fn create_payment_connector(
                 connector_label: connector_label.clone(),
             },
         )?;
-    match mca.connector_webhook_details.clone(){
-        Some(webhook) => {
-            let connector_webhook: api_models::admin::MerchantConnectorWebhookDetails = serde_json::from_value(webhook.peek().to_owned())
+    if let Some(connector_webhook_details) = mca.connector_webhook_details.clone() {
+        let connector_webhook: api_models::admin::MerchantConnectorWebhookDetails =
+            serde_json::from_value(connector_webhook_details.peek().to_owned())
                 .into_report()
                 .attach_printable("failed to deserialize connector_webhook_details")
                 .change_context(errors::ApiErrorResponse::InternalServerError)?;
-            match connector_webhook.merchant_secret{
-                Some(merchant_secret) => {
-                    let config_key = format!("{}_{}", merchant_id, connector_label);
-                    store.insert_config(ConfigNew{
-                        key: config_key,
-                        config: merchant_secret.expose()
-                    })
-                    .await
-                    .attach_printable("attachment")
-                    .change_context(errors::ApiErrorResponse::InternalServerError)?;
-                },
-                None => {}
-            }
-        },
-        None => {},
+        if let Some(merchant_secret) = connector_webhook.merchant_secret {
+            let config_key = format!("{}_{}", merchant_id, connector_label);
+            store
+                .insert_config(ConfigNew {
+                    key: config_key,
+                    config: merchant_secret.expose(),
+                })
+                .await
+                .attach_printable("attachment")
+                .change_context(errors::ApiErrorResponse::InternalServerError)?;
+        }
     };
 
     metrics::MCA_CREATE.add(
@@ -629,15 +627,16 @@ pub async fn update_payment_connector(
         payment_methods_enabled,
         metadata: req.metadata,
         frm_configs,
-        connector_webhook_details: match req.connector_webhook_details{
+        connector_webhook_details: match req.connector_webhook_details {
             Some(connector_webhook_details) => {
-                Encode::<api_models::admin::MerchantConnectorWebhookDetails>::encode_to_value(&connector_webhook_details)
-                    .change_context(errors::ApiErrorResponse::InternalServerError)
-                    .map(|value| Some(Secret::new(value)))?
-            },
+                Encode::<api_models::admin::MerchantConnectorWebhookDetails>::encode_to_value(
+                    &connector_webhook_details,
+                )
+                .change_context(errors::ApiErrorResponse::InternalServerError)
+                .map(|value| Some(Secret::new(value)))?
+            }
             None => None,
-        }
-            
+        },
     };
 
     let updated_mca = db
@@ -647,27 +646,27 @@ pub async fn update_payment_connector(
         .attach_printable_lazy(|| {
             format!("Failed while updating MerchantConnectorAccount: id: {merchant_connector_id}")
         })?;
-    match updated_mca.connector_webhook_details.clone(){
-        Some(webhook) => {
-            let connector_webhook: api_models::admin::MerchantConnectorWebhookDetails = serde_json::from_value(webhook.peek().to_owned())
+
+    if let Some(connector_webhook_details) = updated_mca.connector_webhook_details.clone() {
+        let connector_webhook: api_models::admin::MerchantConnectorWebhookDetails =
+            serde_json::from_value(connector_webhook_details.peek().to_owned())
                 .into_report()
                 .attach_printable("failed to deserialize connector_webhook_details")
                 .change_context(errors::ApiErrorResponse::InternalServerError)?;
-            match connector_webhook.merchant_secret{
-                Some(merchant_secret) => {
-                    let config_key = format!("{}_{}", merchant_id, updated_mca.connector_label.clone());
-                    db.update_config_cached( &config_key,  storage_models::configs::ConfigUpdate::Update {
-                            config: Some(merchant_secret.expose())
-                        })
-                    .await
-                    .attach_printable("attachment")
-                    .change_context(errors::ApiErrorResponse::InternalServerError)?;
+        if let Some(merchant_secret) = connector_webhook.merchant_secret {
+            let config_key = format!("{}_{}", merchant_id, updated_mca.connector_label.clone());
+            db.update_config_cached(
+                &config_key,
+                storage_models::configs::ConfigUpdate::Update {
+                    config: Some(merchant_secret.expose()),
                 },
-                None => {}
-            }
-        },
-        None => {},
+            )
+            .await
+            .attach_printable("attachment")
+            .change_context(errors::ApiErrorResponse::InternalServerError)?;
+        }
     };
+
     let response = updated_mca.try_into()?;
 
     Ok(service_api::ApplicationResponse::Json(response))
@@ -684,7 +683,10 @@ pub async fn delete_payment_connector(
         .to_not_found_response(errors::ApiErrorResponse::MerchantAccountNotFound)?;
 
     let mca = db
-        .find_by_merchant_connector_account_merchant_id_merchant_connector_id(&merchant_id, &merchant_connector_id)
+        .find_by_merchant_connector_account_merchant_id_merchant_connector_id(
+            &merchant_id,
+            &merchant_connector_id,
+        )
         .await
         .to_not_found_response(errors::ApiErrorResponse::MerchantConnectorAccountNotFound {
             id: merchant_connector_id.clone(),
@@ -701,24 +703,19 @@ pub async fn delete_payment_connector(
         })?;
 
     if is_deleted {
-        match mca.connector_webhook_details.clone(){
-            Some(webhook) => {
-                let connector_webhook: api_models::admin::MerchantConnectorWebhookDetails = serde_json::from_value(webhook.peek().to_owned())
+        if let Some(connector_webhook_details) = mca.connector_webhook_details.clone() {
+            let connector_webhook: api_models::admin::MerchantConnectorWebhookDetails =
+                serde_json::from_value(connector_webhook_details.peek().to_owned())
                     .into_report()
                     .attach_printable("failed to deserialize connector_webhook_details")
                     .change_context(errors::ApiErrorResponse::InternalServerError)?;
-                match connector_webhook.merchant_secret{
-                    Some(_merchant_secret) => {
-                        let config_key = format!("{}_{}", merchant_id, mca.connector_label.clone());
-                        db.delete_config_by_key( &config_key)
-                        .await
-                        .attach_printable("attachment")
-                        .change_context(errors::ApiErrorResponse::InternalServerError)?;
-                    },
-                    None => {}
-                }
-            },
-            None => {},
+            if let Some(_merchant_secret) = connector_webhook.merchant_secret {
+                let config_key = format!("{}_{}", merchant_id, mca.connector_label.clone());
+                db.delete_config_by_key(&config_key)
+                    .await
+                    .attach_printable("attachment")
+                    .change_context(errors::ApiErrorResponse::InternalServerError)?;
+            }
         };
     }
     let response = api::MerchantConnectorDeleteResponse {
